backend/requirements.txt
flask
psycopg2-binary
flask-cors
openai>=1.0.0
requests
